+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `@all()` decorator to specify all matching targets. Equivalent to`@inject(identifier, { all: true})`
 - Add `Container.create` and factory `context.create` to directly instantiate and inject an arbitrary constructor.
+- Add `bind(Constructor).toSelf()`.
 
 ## Refactors